# models/author.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// Author author
//
// swagger:model Author
type Author struct {

	// author id
	// Example: 1
	AuthorID int64 `json:"author_id,omitempty"`

	// name
	// Example: Dmitrii
	// Required: true
	Name *string `json:"name"`

	// surname
	// Example: Urievich
	// Required: true
	Surname *string `json:"surname"`
}

// Validate validates this author
func (m *Author) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateName(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateSurname(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Author) validateName(formats strfmt.Registry) error {

	if err := validate.Required("name", "body", m.Name); err != nil {
		return err
	}

	return nil
}

func (m *Author) validateSurname(formats strfmt.Registry) error {

	if err := validate.Required("surname", "body", m.Surname); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this author based on context it is used
func (m *Author) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Author) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Author) UnmarshalBinary(b []byte) error {
	var res Author
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}