# swagger.yaml
swagger: "2.0"
info:
  title: Book authors
  description: "This is server about authors"
  version: 1.0.0
host: localhost:8080
schemes: 
  - http
basePath: /v1
paths:
  /authors:
    get:
      summary: "get all authors"
      tags:
         - Authors
      operationId: getAllAuthors
      produces:
       - application/json
      responses:
        "200":
          description: "successful response"
          schema: 
             $ref: "#/definitions/Authors"
        "default":
           description: "Все нестандартное"
           schema: 
              $ref: "#/definitions/Error"
    post:
      tags:
         - Authors
      summary: "Add a new author to the store"
      operationId: "addAuthor"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Author that needed to be in the list"
        required: true
        schema:
          $ref: "#/definitions/Author"
      responses:
        "201":
          description: "Created"
          schema:
            properties: 
              id:
                type: integer
        "405":
          description: "Invalid input"
          schema:
            properties: 
              message:
                type: string
  /authors/{authorId}:
    get:
      tags:
         - Authors
      summary: "Find author by ID"
      description: "Returns a single author"
      operationId: "getAuthorById"
      produces:
      - "application/json"
      parameters:
      - name: "authorId"
        in: "path"
        description: "ID of author to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Author"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Author not found"
    delete:
      tags:
         - Authors
      summary: "Deletes an author"
      operationId: "deleteAuthor"
      produces:
      - "application/json"
      parameters:
      - name: "authorId"
        in: "path"
        description: "ID author to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Author not found"

definitions: 
    Author:
      type: object
      required:
      - name
      - surname
      properties:
        author_id:
          type: integer
          example: 1
        name:
          type: string
          example: Dmitrii
        surname: 
          type : string
          example: Urievich
    Authors:
      type: array
      items: 
        $ref: "#/definitions/Author"
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message: 
          type: string
          
          
    
    